Add a test environment
This commit adds a container with Lua 5.2 and the `busted` library
installed.

It also adds a Makefile target to extract only the Lua code from the
PICO-8 cart, allowing the busted spec to load the game's Lua code.

The test target extract the Lua code from the cart, and runs the busted
specs in the Docker container.